+++ apps/server/src/modules/user-login-migration/api/uc/toggle-user-login-migration.uc.spec.ts
@@ -12,8 +12,8 @@ import { UserLoginMigrationDO } from '@shared/domain/domainobject';
 import { Permission } from '@shared/domain/interface';
 import { setupEntities } from '@testing/database';
 import { userLoginMigrationDOFactory } from '@testing/factory/domainobject';
-import { UserLoginMigrationNotFoundLoggableException } from '../loggable';
-import { UserLoginMigrationService } from '../service';
+import { UserLoginMigrationNotFoundLoggableException } from '../../domain/loggable';
+import { UserLoginMigrationService } from '../../domain/service';
 import { ToggleUserLoginMigrationUc } from './toggle-user-login-migration.uc';
 
 describe(ToggleUserLoginMigrationUc.name, () => {
