+++ apps/server/src/modules/user-login-migration/index.ts
@@ -1,3 +1,3 @@
-export * from './user-login-migration.module';
-export * from './service';
+export { MigrationCheckService, UserLoginMigrationService, UserMigrationService } from './domain/service';
 export { UserLoginMigrationConfig } from './user-login-migration.config';
+export * from './user-login-migration.module';
