--- apps/server/src/modules/user-login-migration/uc/index.ts
@@ -1,8 +1,0 @@
-export * from './dto';
-export * from './user-login-migration.uc';
-export * from './start-user-login-migration.uc';
-export * from './toggle-user-login-migration.uc';
-export * from './restart-user-login-migration.uc';
-export * from './close-user-login-migration.uc';
-export * from './user-login-migration-rollback.uc';
-export { CloseMigrationWizardUc } from './close-migration-wizard.uc';
