+++ apps/server/src/modules/user-login-migration/user-login-migration-api.module.ts
@@ -7,8 +7,6 @@ import { ProvisioningModule } from '@modules/provisioning';
 import { UserModule } from '@modules/user';
 import { ImportUserModule } from '@modules/user-import';
 import { Module } from '@nestjs/common';
-import { UserLoginMigrationRollbackController } from './api/user-login-migration-rollback.controller';
-import { UserLoginMigrationController } from './api/user-login-migration.controller';
 import {
 	CloseMigrationWizardUc,
 	CloseUserLoginMigrationUc,
@@ -17,7 +15,9 @@ import {
 	ToggleUserLoginMigrationUc,
 	UserLoginMigrationRollbackUc,
 	UserLoginMigrationUc,
-} from './uc';
+} from './api/uc';
+import { UserLoginMigrationRollbackController } from './api/user-login-migration-rollback.controller';
+import { UserLoginMigrationController } from './api/user-login-migration.controller';
 import { UserLoginMigrationModule } from './user-login-migration.module';
 
 @Module({
